+++ src/Application/DownloadController.php
@@ -82,10 +82,16 @@ class DownloadController extends ControllerAbstract
 					);
 				}
 
-				$youtube_provider = \YoutubeDownloader\Provider\Youtube\Provider::createFromConfigAndToolkit(
-					$config,
-					$toolkit
-				);
+				$options = [
+					'decipher_signature' => $config->get('enable_youtube_decipher_signature')
+				];
+
+				if ( $config->get('multipleIPs') === true)
+				{
+					$options['use_ip'] = $toolkit->getRandomIp($config);
+				}
+
+				$youtube_provider = \YoutubeDownloader\Provider\Youtube\Provider::createFromOptions($options);
 
 				if ( $youtube_provider instanceOf \YoutubeDownloader\Cache\CacheAware )
 				{